@@ -18,38 +18,51 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.0.3",
-    "@angular/cdk": "^12.0.3",
-    "@angular/common": "~12.0.3",
-    "@angular/compiler": "~12.0.3",
-    "@angular/core": "~12.0.3",
-    "@angular/flex-layout": "^12.0.0-beta.34",
-    "@angular/forms": "~12.0.3",
-    "@angular/material": "^12.0.3",
-    "@angular/platform-browser": "~12.0.3",
-    "@angular/platform-browser-dynamic": "~12.0.3",
-    "@angular/router": "~12.0.3",
-    "@azure/msal-angular": "^2.0.0",
-    "@azure/msal-browser": "^2.14.2",
-    "@microsoft/applicationinsights-web": "^2.6.2",
-    "@ngneat/until-destroy": "^8.0.4",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.1.0",
+    "@angular/animations": "~13.2.0",
+    "@angular/cdk": "^13.2.1",
+    "@angular/common": "~13.2.0",
+    "@angular/compiler": "~13.2.0",
+    "@angular/core": "~13.2.0",
+    "@angular/flex-layout": "^13.0.0-beta.38",
+    "@angular/forms": "~13.2.0",
+    "@angular/material": "^13.2.1",
+    "@angular/platform-browser": "~13.2.0",
+    "@angular/platform-browser-dynamic": "~13.2.0",
+    "@angular/router": "~13.2.0",
+    "@azure/msal-angular": "^2.1.0",
+    "@azure/msal-browser": "^2.21.0",
+    "@microsoft/applicationinsights-web": "^2.7.3",
+    "@ngneat/until-destroy": "^9.0.0",
+    "rxjs": "~7.5.0",
+    "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.0.3",
-    "@angular/cli": "~12.0.3",
-    "@angular/compiler-cli": "~12.0.3",
-    "@types/jasmine": "~3.6.0",
+    "@angular-devkit/build-angular": "~13.2.2",
+    "@angular-eslint/builder": "13.0.1",
+    "@angular-eslint/eslint-plugin": "13.0.1",
+    "@angular-eslint/eslint-plugin-template": "13.0.1",
+    "@angular-eslint/schematics": "13.0.1",
+    "@angular-eslint/template-parser": "13.0.1",
+    "@angular/cli": "~13.2.2",
+    "@angular/compiler-cli": "~13.2.0",
+    "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
-    "jasmine-core": "~3.7.0",
+    "@typescript-eslint/eslint-plugin": "5.3.0",
+    "@typescript-eslint/parser": "5.3.0",
+    "concurrently": "^7.0.0",
+    "dev-error-reporter": "^1.0.8",
+    "eslint": "^8.2.0",
+    "jasmine-core": "~4.0.0",
+    "json-concat": "0.0.1",
+    "json-server": "^0.17.0",
     "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
+    "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
     "karma-junit-reporter": "^2.0.1",
-    "typescript": "~4.2.3"
+    "onchange": "^7.1.0",
+    "typescript": "~4.5.2"
   }
 }
